Step 6 - turns on restrictions
