+++ packages/server/src/data/channel-memberships-data.js
@@ -6,8 +6,6 @@ const { ChannelMembershipModel } = require('../models-abstract');
 const { CHANNEL_MEMBERSHIP_ROLES } = require('../models/enums/channel-membership-roles');
 const { CHANNEL_MEMBERSHIP_TYPE } = require('../models/enums/channel-membership-type');
 
-const structures = require('../structures');
-
 /**
  * @implements {Layers.DataSourceLayer.DataSource}
  * @property {ChannelMembershipModel} ChannelMembershipModel
@@ -92,10 +90,10 @@ class ChannelMembershipsData extends DataBase {
    */
   async _validate({ channelId, agentId }) {
     if (!await this.models.Channel.findByPk(channelId)) {
-      return Promise.reject(new Error(`Channel "${channelId}" is not found.`));
+      return Promise.reject(new Error(`Channel "${ channelId }" is not found.`));
     }
     if (!await this.models.Agent.findByPk(agentId)) {
-      return Promise.reject(new Error(`Agent "${agentId}" is not found.`));
+      return Promise.reject(new Error(`Agent "${ agentId }" is not found.`));
     }
     return true;
   }
@@ -199,19 +197,8 @@ class ChannelMembershipsData extends DataBase {
    * @return {Layers.AbstractModelsLayer.restrictions}
    */
   _getDefaultRestrictions(options = {}) {
-    const currentAgentId = _.get(options, 'currentAgentId');
-
-    const withinPublicChannels = currentAgentId
-      ? null
-      : structures.withinPublicChannelsByChannelId();
-
-    const withinPublicAndJoinedChannels = currentAgentId
-      ? structures.withinPublicAndJoinedChannelsByChannelId(currentAgentId)
-      : null;
-
     return {
-      withinPublicChannels,
-      withinPublicAndJoinedChannels,
+      ...super._getDefaultRestrictionsForIncludedIntoChannel(options),
     };
   }
 }
